pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognitive"
version = "0.1.0"
description = "Cognitive Modules CLI - Structured LLM task runner with schema validation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lucio", email = "lucio@example.com"}
]
keywords = ["llm", "cognitive", "modules", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "jsonschema>=4.17.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
ollama = ["requests>=2.28.0"]
all = ["openai>=1.0.0", "anthropic>=0.18.0", "requests>=2.28.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
cog = "cognitive.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/cognitive"
Documentation = "https://github.com/your-org/cognitive#readme"
Repository = "https://github.com/your-org/cognitive"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.md", "*.txt"]
